requirements.txt
# Requerimientos con versiones fijadas (Python 3.8.10)
PyQt6==6.7.1
openai==1.102.0
python-docx==1.1.2
python-pptx==1.0.2
PyMuPDF==1.24.11
fpdf2==2.8.3
Pillow==10.4.0
pytesseract==0.3.13

# Dependencias explícitas fijadas para instalación reproducible
PyQt6-Qt6==6.7.3
PyQt6-sip==13.8.0
anyio==4.5.2
distro==1.9.0
httpx==0.28.1
jiter==0.9.1
pydantic==2.10.6
sniffio==1.3.1
tqdm==4.67.1
typing-extensions==4.13.2
lxml==6.0.1
XlsxWriter==3.2.5
defusedxml==0.7.1
fonttools==4.57.0
packaging==25.0
idna==3.10
exceptiongroup==1.3.0
certifi==2025.8.3
httpcore==1.0.9
h11==0.16.0
annotated-types==0.7.0
pydantic-core==2.27.2
colorama==0.4.6
